+++ src/main/java/com/backend/makeUrTasks/makeUrTasks/Interfaces/TaskModelInterface.java
@@ -6,16 +6,16 @@ import java.util.ArrayList;
 
 public interface TaskModelInterface {
 
-  public AbstractTask findByTitle(String title, float userId);
+  public AbstractTask findByTitle(String title, Integer userId);
 
-  public AbstractTask findById(float taskId, float userId);
+  public AbstractTask findById(Integer taskId, Integer userId);
 
-  public ArrayList<AbstractTask> find(float userId, int page);
+  public ArrayList<AbstractTask> find(Integer userId, int page);
 
-  public AbstractTask create(String title, float userId, String description);
+  public AbstractTask create(String title, Integer userId, String description);
 
-  public void delete(float taskId);
+  public void delete(Integer taskId);
 
-  public AbstractTask edit(String title, String description, float taskId);
+  public AbstractTask edit(String title, String description, Integer taskId);
 
 }
